+++ Cargo.toml
@@ -4,7 +4,6 @@ members = [
   "helix-view",
   "helix-term",
   "helix-tui",
-  "helix-syntax",
   "helix-lsp",
   "helix-dap",
   "xtask",
@@ -13,10 +12,6 @@ members = [
 default-members = [
   "helix-term"
 ]
-
-# Build helix-syntax in release mode to make the code path faster in development.
-# [profile.dev.package."helix-syntax"]
-# opt-level = 3
 
 [profile.dev]
 split-debuginfo = "unpacked"
