+++ helix-term/Cargo.toml
@@ -66,5 +66,9 @@ grep-searcher = "0.1.8"
 # Remove once retain_mut lands in stable rust
 retain_mut = "0.1.7"
 
+# compiling grammars
+cc = { version = "1" }
+threadpool = { version = "1.0" }
+
 [target.'cfg(not(windows))'.dependencies]  # https://github.com/vorner/signal-hook/issues/100
 signal-hook-tokio = { version = "0.3", features = ["futures-v0_3"] }
